% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_package.R
\docType{package}
\name{cSEM.DGP-package}
\alias{cSEM.DGP}
\alias{cSEM.DGP-package}
\title{Generate data for structural equation models}
\description{
Generate data for structural equation models including up to 8 constructs. Generation
is based parameter values given in \href{http://lavaan.ugent.be/tutorial/syntax1.html}{lavaan model syntax}.
In addition to supplying numeric values, variable values for parameters are allowed.
To achieve this, the package makes use of \href{http://lavaan.ugent.be/}{lavaan}'s
labeling capabilities.
Users may replace a given parameter in, i.e. the structural model
by a symbolic name and assign a vector of values to that name. These values will be used
to generate data for all possible combinations of these values with the remaining fixed parameters.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/M-E-Rademaker/cSEM.DGP}
  \item Report bugs at \url{https://github.com/M-E-Rademaker/cSEM.DGP/issues}
}

}
\author{
\strong{Maintainer}: Manuel Rademaker \email{manuel.rademaker@uni-wuerzburg.de} (\href{https://orcid.org/0000-0002-8902-3561}{ORCID})

Authors:
\itemize{
  \item Tamara Schamberger \email{tamara.schamberger@uni-wuerzburg.de} (\href{https://orcid.org/0000-0002-7845-784X}{ORCID})
}

}
\keyword{internal}
