% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatecSEMModel.r
\name{generatecSEMModel}
\alias{generatecSEMModel}
\title{Internal: generate cSEMModels}
\usage{
generatecSEMModel(.model, ...)
}
\arguments{
\item{.model}{A model in \link[lavaan:model.syntax]{lavaan model syntax} possibly
containing labels.}

\item{...}{\code{"name" = values} pairs. \code{"name"} is a character string giving the
label used for the parameter of interest. \code{values} is a numeric vector of
values to use for the paramter given by \code{"name"}.}
}
\value{
A list of \link[cSEM:csem_model]{cSEMModel}s
}
\description{
Generate all possible \link[cSEM:csem_model]{cSEMModel}s.
}
\keyword{internal}
