% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSigma.R
\name{generateSigma}
\alias{generateSigma}
\title{Generate indicator correlation matrix}
\usage{
generateSigma(
 .model                    = NULL,
 .handle_negative_definite = c("stop", "ignore")
 )
}
\arguments{
\item{.model}{A model in \link[lavaan:model.syntax]{lavaan model syntax}}

\item{.handle_negative_definite}{Character string. How should negative definite
Sigma matrices be handeled? One of \code{"stop"} or \code{"ignore"} in which case
an \code{NA} is produced. Defaults to \code{"stop"}.}
}
\value{
A K by K matrix of indicator correlations. K is the number of indicators
}
\description{
Generates a population indicator correlation matrix for a given
structural equation model based on the parameter given by the user.
}
\examples{
model <- "
# Structural model
eta2 ~ 0.6*eta1
eta3 ~ 0.4*eta1 + 0.35*eta2

# Measurement model
eta1 =~ 0.8*y11 + 0.9*y12 + 0.8*y13
eta2 =~ 0.7*y21 + 0.7*y22 + 0.9*y23
eta3 =~ 0.9*y31 + 0.8*y32 + 0.7*y33
"

Sigma <- generateSigma(model)
Sigma
}
