% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateConstructCor.R
\name{generateConstructCor}
\alias{generateConstructCor}
\title{Compute construct correlation matrix}
\usage{
generateConstructCor(path_matrix = path_matrix, gamma = NULL)
}
\arguments{
\item{gamma}{A (8 x 8) matrix containing the path coefficients.}

\item{.structural}{A matrix mimicking the structural relationship
between constructs. Up to 8 constructs are supported.}
}
\value{
A (8 x 8) matrix of construct correlations.
}
\description{
Calculate the construct correlation matrix correlation. Currently, there
is a limit of ten constructs.
}
\keyword{internal}
