% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatecSEMModel.r
\name{generatecSEMModel}
\alias{generatecSEMModel}
\title{Generate cSEMModels}
\usage{
generatecSEMModel(.model, ...)
}
\arguments{
\item{.model}{A model in \link[lavaan:model.syntax]{lavaan model syntax} possibly
containing labels.}

\item{...}{\code{"name" = values} pairs. \code{"name"} is a character value giving the
label used for the parameter of interest. \code{values} is a numeric vector of
values to use for the paramter given by \code{"name"}.}
}
\value{
A list of \link[cSEM:csem_model]{cSEMModel}s
}
\description{
Generate all possible \link[cSEM:csem_model]{cSEMModel}s.
}
\examples{
model <- "
# Structural model
eta2 ~ gamma1*eta1
eta3 ~ 0.4*eta1 + 0.35*eta2

# Measurement model
eta1 =~ lambda1*y11 + 0.9*y12 + 0.8*y13
eta2 =~ 0.7*y21 + 0.7*y22 + 0.9*y23
eta3 =~ 0.9*y31 + 0.8*y32 + 0.7*y33
"

Models <- generatecSEMModel(model,
                           "gamma1" = c(0.3, 0.6),
                           "lambda1" = c(0.8, 0.85, 0.9))
Models

}
