% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSigma.R
\name{generateSigma}
\alias{generateSigma}
\title{Generate the indicator correlation matrix}
\usage{
generateSigma(
 .model                    = NULL,
 .handle_negative_definite = c("stop", "drop", "set_NA")
 )
}
\arguments{
\item{.model}{A model generated using \code{\link[=generatecSEMModel]{generatecSEMModel()}}.}

\item{.handle_negative_definite}{Character string. How should negative definite
indicator correlation matrices be handled? One of \code{"stop"}, \code{"drop"} or \code{"set_NA"}
in which case an \code{NA} is produced. Defaults to \code{"stop"}.}
}
\value{
A K by K matrix of indicator correlations. K is the number of indicators
}
\description{
Generates the indicator correlation matrix based on the parameters of a
structural equation model in \link[lavaan:model.syntax]{lavaan model syntax}.
}
\keyword{internal}
